dev (cluster/brokers cli) make brokers a subcomand from insprctl
